// types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      term_types: {
        Row: {
          created_datetime_utc: string
          id: number
          name: string
        }
        Insert: {
          created_datetime_utc?: string
          id?: number
          name: string
        }
        Update: {
          created_datetime_utc?: string
          id?: number
          name?: string
        }
        Relationships: []
      }
      terms: {
        Row: {
          created_datetime_utc: string
          definition: string
          example: string
          id: number
          modified_datetime_utc: string | null
          priority: number
          term: string
          term_type_id: number | null
        }
        Insert: {
          created_datetime_utc?: string
          definition: string
          example: string
          id?: number
          modified_datetime_utc?: string | null
          priority?: number
          term: string
          term_type_id?: number | null
        }
        Update: {
          created_datetime_utc?: string
          definition?: string
          example?: string
          id?: number
          modified_datetime_utc?: string | null
          priority?: number
          term?: string
          term_type_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "terms_term_type_id_fkey"
            columns: ["term_type_id"]
            isOneToOne: false
            referencedRelation: "term_types"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}